Currently depends on a alpha version of `futures-preview`, but with `std::future` now on stable we should get a stable release of futures 0.3 soon. Normally I would say wait on something like this, but #2385 is going to need this change to properly support async ops.