@95th 2019-07-18
Can we temporarily switch to nightly for some time and move the code to async await?

cc @ry