@ry 2019-07-18
> Can we temporarily switch to nightly for some time and move the code to async await?

You can do that in a branch :) 

Not in master. I don't want to open that can of worms.