@95th 2019-07-18
In that case, given the scale of refactoring we're doing at the moment, I think it is not worth it to maintain another branch. We'll do it when async-await is in stable