@afinch7 2019-07-18
Here is a list of dependencies, and their respective statuses.

- [ ] [tokio](https://github.com/tokio-rs/tokio) (`tokio-executor`, `tokio-fs`, `tokio-io`, `tokio-threadpool`)       waiting on alpha release tokio-rs/tokio#1201 (using git for now)
  - [ ] [futures](https://github.com/rust-lang-nursery/futures-rs) - [waiting on stable release](https://github.com/rust-lang-nursery/futures-rs/milestone/2) (using preview/alpha for now)
  - [ ] [async-await](https://github.com/rust-lang/rust/issues/50547) - waiting on stabilization (using nightly until 1.38.0 release) rust-lang/rust#62149
  

Dependent on tokio:
- [ ] [tokio-process](https://github.com/alexcrichton/tokio-process) - working on a patch 
    alexcrichton/tokio-process#59
- [ ] [hyper](https://github.com/hyperium/hyper) - waiting on alpha release (using git for now)
- [ ] [hyper-rustls](https://github.com/ctz/hyper-rustls) - waiting on patch to be started(dependent on 
    others)
  - [ ] [tokio-rustls](https://github.com/quininer/tokio-rustls) - working on a patch


I think that is everything let me know if there is something I missed. Also thoughts on [runtime](https://github.com/rustasync/runtime)?