@afinch7 2019-07-18
I manged to get `deno_tcp` working, so here are some results so far. I picked the median of 3 runs for both. Don't mind the terrible tail latency. I think that has more to do with my hardware than anything.

This branch:
```
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.91ms    5.76ms 102.88ms   98.33%
    Req/Sec    14.77k     1.69k   16.43k    96.00%
  293777 requests in 10.00s, 14.29MB read
Requests/sec:  29369.49
Transfer/sec:      1.43MB
```

Master:
```
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.91ms    5.81ms 104.34ms   98.42%
    Req/Sec    14.55k     1.67k   16.07k    96.00%
  289246 requests in 10.00s, 14.07MB read
Requests/sec:  28920.91
Transfer/sec:      1.41MB
```

Within margin in error, but still about ~2-3% in favor of this branch.