@95th 2019-08-22
> Here is a list of dependencies, and their respective statuses.

Think most of the stuff can be checked except hyper-rustls (Perhaps use hyper-tls)?