@afinch7 2019-08-22
> > Here is a list of dependencies, and their respective statuses.
> 
> Think most of the stuff can be checked except hyper-rustls (Perhaps use hyper-tls)?

Yes I think most of the dependencies are ready, but async/await has been delayed to 1.39.0 that's really the only thing I'm still waiting on here.