@ry 2019-10-02
Old PR, closing. 

We will move to the new futures API, but ideally there would be a stable release of Tokio using this new future API first. We need Tokio to make a stable release using the new Futures API first.